requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
authlib
python-dotenv
